+++ classes/Conf/class.xoctConf.php
@@ -49,6 +49,7 @@ class xoctConf extends ActiveRecord {
 	const F_WORKFLOW_PARAMETERS = 'workflow_parameters';
 	const F_AUDIO_ALLOWED = 'audio_allowed';
 	const F_CREATE_SCHEDULED_ALLOWED = 'create_scheduled_allowed';
+	const F_STUDIO_ALLOWED = 'oc_studio_allowed';
 	const F_VIDEO_PORTAL_LINK = 'video_portal_link';
 	const F_VIDEO_PORTAL_TITLE = 'video_portal_title';
 	const F_ENABLE_LIVE_STREAMS = 'enable_live_streams';